Fix ordering filter and improve query performance
Updated the parameters used in the GetAllListForMoveOrderPagination method in OrderingController.cs, the dateTo parameter was mistakenly used twice instead of using dateFrom and dateTo.

In the OrderingRepository.cs, the backend query was refactored to consolidate the individual filters into one WHERE statement to enhance the query performance. Parsing to DateTime was moved out of the query itself to not execute it multiple times during query runtime. This greatly improves query performance when working with large data sets.