"Updated conditional check and DB connection in OrderingRepository and Startup.
In OrderingRepository.cs, the checking condition has changed from 'OrderNo' to 'OrderNoPKey' and IsPrint property is now always set to true instead of switching between true and false. The changes ensure that the check is based on primary key and that print flag is always true leading to better object identification and use-cases, respectively.

In Startup.cs, the ConnectionStringUpdated has been changed to 'LiveConnectionDevelopment' from 'LiveConnection' for debugging and development purposes.

In ReportRepository.cs, variable name for PreparedDate has been removed to improve readability and keep the code clean."

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>