Modify MoveOrders query and save changes to the database
Changes were made to OrderingRepository.cs, specifically targeting the "MoveOrders" query and how changes are saved. The "IsTransact" field was cleaned up to remove extraneous data. In the loop for order numbers, the check was adjusted to look at "OrderNo" instead of "OrderNoPKey", reflecting a more accurate traversal of Order instances. Finally, a database save operation was placed in the loop to immediately apply any changes made to the fields of the Order objects.

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>