Refactor MultiplePrintingForMOS method in OrderingRepository
The MultiplePrintingForMOS method in OrderingRepository was changed to address grouping of orders. We implemented this change to improve the readability and efficiency in the printing of multiple orders. Now, it groups orders by OrderNumber and maps them to a new MultiplePrintingDTO object. The DTO was also created and added to the directory, and the function signature in IOrdering.cs interface was updated accordingly. This should enhance the overall printing process.

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>