Add ApprovedMoveOrderReport to ReportRepository and ReportController
A new report called ApprovedMoveOrderReport was added to the report repository interface and its implementations. The new API endpoint was also created in ReportController ("ApprovedMoveOrdersReports") to handle HTTP GET requests for this report. This was done to facilitate fetching of approved move orders within a specified date range, a feature which was not previously available. Operations for searching and pagination in OrderingController were also amended for this new functionality.

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>