Refactor UpdatePrintStatus and expand searching parameters
Refactored the UpdatePrintStatus function to accept a MoveOrder object instead of an int array. This adjustment simplifies the code and eliminates the need for a loop that iterated through the array, improving performance.

Additionally, expanded the search parameters in the ReceivingRepository to include not just PO_Number, but also ItemDescription and ItemCode. This change will improve the system's search functionality, making it more flexible and user-friendly.

Also made minor modifications on Startup.cs to change LiveConnectionDevelopment to LiveConnection. Changed approved MoveOrderReport's return object type from MoveOrderDto to MoveOrderReport to accommodate newly added field 'DeliveryStatus' in MoveOrderReport class. Various interface and controller changes have been made to align with these modifications.

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>