"Refactored class names and extended search functionality"
This commit does several things. Firstly, it renames the class "AddNewProspect" to the more appropriate "ExportReports" in ExportMoveOrderReport.cs, alignifying the class name with the file's function.

In OrderingRepository.cs, the search functionality was augmented to include "FarmName" in addition to "OrderNo" providing users additional flexibility when making searches.

In ExportConsolidatedReport.cs, a new file was created for feature Exports which corresponds to actions available for manipulating exports and reports in our application.

Several changes have been made to the RawMaterialInventory.cs file to modify the output of the inventory report to the preferred format.

The Startup.cs file has been modified to use the live connection string rather than the development connection string.

In ConsolidatedReport.cs, several property declarations have been modified and new ones have been added to capture more relevant information for reporting purposes.

Eventually, the ReportRepository.cs file has been modified to implement the changes made to the ConsolidatedReport model.

These changes were necessary to improve the overall functionality, readability and maintainability of the codebase.

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>