Change connection string and add UOM to ReportRepository
Updated connection string in Startup.cs to use the "LiveConnection" instead of "LiveConnectionDevelopment". This was done to make the application connect to production database.

Also, the 'UOM' field was added to the ReportRepository in the DATA ACCESS LAYER. This was required to provide more detailed item information in the generated reports.

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>