Update customer validation and report retrieval methods
Updated the CustomerCodeExist method in CustomerRepository.cs by adding an extra parameter (customerName) to enhance data validation. This change will prevent potential data conflict by checking both customerCode and customerName when validating the existence of a customer.

Also, modified the ReportRepository.cs to asynchronously fetch MoveOrders and apply the filters on the fetched data to improve application performance. This approach will reduce the load time of the reports generated when fetching MoveOrders.

Signed-off-by: aldrinvega <aldrinmundovega@gmail.com>